Basic framework of debugger
Summary:
- set up debugger in prepack and connect
- set up proxy to communicate with client
- currently turned debugger off with a flag
