+++ .gitignore
@@ -10,3 +10,4 @@ coverage*/
 facebook/test
 facebook/tutorial
 prepack.min.js
+src/debugger/proxy/*.pyc
