+++ src/debugger/proxy/main.py
@@ -1,0 +1,8 @@
+from session import Session
+
+def main():
+    s = Session()
+    s.serve()
+
+if __name__ == "__main__":
+    main()
