+++ src/environment.js
@@ -1253,7 +1253,7 @@ export class LexicalEnvironment {
   }
 
   evaluate(ast: BabelNode, strictCode: boolean, metadata?: any): Value | Reference {
-    if (this.realm.attachedDebugger && ast.loc) {
+    if (this.realm.enableDebugger && this.realm.attachedDebugger && ast.loc) {
       this.realm.attachedDebugger.checkForActions(ast, this.realm.contextStack);
     }
     let res = this.evaluateAbstract(ast, strictCode, metadata);
